Move tag stat functions to separate file
