+++ src/analizer.ts
@@ -4,6 +4,7 @@ import { getAllUserComments, isToRMod } from './reddit-api';
 import { gammaAvg, karmaAvg } from './stats/avg';
 import { gammaPeak, karmaPeak } from './stats/peak';
 import { subredditGamma, subredditKarma } from './stats/subreddits';
+import { getCountTag, getModTag, getSpecialTags } from './stats/tags';
 import { formatGamma, typeGamma, formatKarma, typeKarma } from './stats/type';
 import { CountTag, specialTags, countTags, Tag } from './tags';
 import Transcription from './transcription';
@@ -18,62 +19,6 @@ export function isComment(comment: Comment): boolean {
     // Has one of the bot keywords
     /\b(done|(un)?claim(ing)?)\b/.test(comment.body)
   );
-}
-
-/**
- * Gets the matching count tag for the transcriptions.
- * @param transcriptions The transcriptions to analize.
- */
-export function getCountTag(transcriptions: Transcription[]): CountTag {
-  const count = transcriptions.length;
-
-  // From the highest tag downwards, search for the first match
-  for (let i = countTags.length - 1; i >= 0; i -= 1) {
-    if (count >= countTags[i].lowerBound) {
-      return countTags[i];
-    }
-  }
-
-  throw new Error(`No count tag found for count ${count}`);
-}
-
-/**
- * Returns the mod tag if the user is a mod or null otherwise.
- * @param userName The username to get the mod tag for.
- */
-export async function getModTag(userName: string): Promise<Tag | null> {
-  const isMod = await isToRMod(userName);
-
-  if (isMod) {
-    return specialTags.mod;
-  }
-
-  return null;
-}
-
-/**
- * Returns the 100/24h tag if the user has completed it, else null.
- * @param transcriptions The transcriptions to analyze.
- */
-export function getTwentyFourTag(transcriptions: Transcription[]): Tag | null {
-  const dayPeak = gammaPeak(transcriptions, 24 * 60 * 60); // 24h
-
-  if (dayPeak.peak >= 100) {
-    return specialTags.twentyFour;
-  }
-
-  return null;
-}
-
-/**
- * Returns all special tags (except the mod tag) for the given user.
- * @param userName The user to check the special tags for.
- * @param transcriptions The transcriptions of the user.
- */
-export function getSpecialTags(userName: string, transcriptions: Transcription[]): Tag[] {
-  const twentyFourTag = getTwentyFourTag(transcriptions);
-
-  return [twentyFourTag].filter((tag) => tag !== null) as Tag[];
 }
 
 type TranscriptionAmount = {
